weather-microservices/auth-service/requirements.txt
Flask==3.0.3
PyJWT==2.9.0
prometheus-flask-exporter==0.23.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-flask==0.48b0
python-json-logger==2.0.7
pytest==8.3.2
